Add filter for "connection reset by peer", Rename previous filter, Be less verbose
This commit renames the InitialHandlerFilter class to IHConnectedFilter and
introduces a new filter named IHResetByPeerFilter. Because of the new class,
the old name would have been confusing. Of course, there is an unit test for
the new class included out of the box. Additionally, the default configuration
is updated here. For existing setups, this shouldn't make any difference.

Additionally, this commit introduces a new setting called "debug mode", which
turns on the filter registration messages, although the default is off. This
has been done because these messages are mostly harm-/useless to normal users
and the only case where they could be of use is when debugging.

I must only make one change in a commit.
I must only make one change in a commit.